Adding initial Savant homebrew class draft document
This is the latest draft I have for the Savant homebrewed class from laserllama on GM Binder. As of right now, it only has the main class and two subclasses (Archaeologist and Physician) and I'm not sure if they're up to date, code-wise. Code has been submitted to MPMB discord for assistance with checking it for errors. Will be updating this file as more content is added.